Reorganized quadcopter radio into class
uart.h/uart.c used to be plain C source. To make the interface work into the communication layer, I reorganized the C code into a class (RadioUART) that inherits from Radio. That way, the quadcopter radio can also be used with the RadioCommunication class, the same way the remote uses RadioLinux class.

Also revamped the Makefiles so that rules are automatically generated based on a single list of object names. This should make it much easier to modify as source files are added.

The code that is shared by both the quadcopter and the remote is now in /common/. This includes radio communication code, which is meant to be used on both sides of communication.