RadioConnection::connect()
Added a function RadioConnection::connect() which waits for an acknowledgement from the other side before returning. This is to allow synchronization for the beginning of communication.

Wrote a test which waits for a connection, then alternates between sending and receiving Packets, to be run on both sides at the same time.