More orientation
Trying to figure out the best way to get/represent orientation. Right now, in test_orientation, 0 deg means "down" and orientation is divided into rotation about x-axis and rotation about y-axis.

The computation forces the accelerometer to be close to the upright position; other orientations cause measurements in the tilted axis to be very inaccurate. Might just use this way anyway (since it's easy) and force the accelerometer to be physically upright.