Tested PWM Dithering
Fixed a few bugs with the dithering. There were too many sub-quantization levels for the PWM frequency that the motor controllers use - at 50Hz, 8 sub-levels means that the dithering cycle is at least 6.25Hz, so each cycle was so slow it was even audible. Decreased it to 4 sub-levels; it might be good to go even less than that depending on testing with actual stabilization.

Tried to reproduce an I2C communication bug that was happening when trying to read from Accelerometer or Gyroscope (ioctl failure...), but I can't reproduce it anymore. Tried feeding different voltages to the power supply which I thought was related to the issue (hi battery voltage seemed to be correlated with more failures), but with a voltage-controlled power supply it isn't reproduced. Might just have to ignore the I2C errors in the code.