PWM dithering
Implemented dithering for the PWM signal in the PWM class. To use dithering,
update() must be called on PWM objects periodically. The effectiveness of the
dither depends on the frequency of the update.

This hasn't been tested yet; likely to be a lot of bugs. Wrote a test program
to run on the quadcopter.