Two-Staged PID; Timed Synchronous Update
Redesigned the update routine so that it is controlled by a timer. The user of the Drive class specifies an update rate, and then a timer signals a separate thread which runs the update routine. Hopefully this will help with reliability and also be more synchronous with the sensor update rate. Stopping and restarting the timer hasn't been tested, though it's probably not a very useful API.

Changed the PID control to be a two-staged ("cascaded") PID control. In this setup, the output from one PID is used as the input for the 2nd PID (PID in series), and the 2nd PID output goes to the motor speed. The 1st PID takes the angle as input, while the 2nd PID takes angular *rate* as input (i.e. gyroscope). In this way, the control loop acts as if trying to achieve the appropriate angular rate instead of directly the desired angle. Credit for idea to Oscar Liang (blog.oscarliang.net/quadcopter-pid-explained-tuning/)